refactor(crypto): reorganize token hashing and signing functionality
