+++ api/models/inviteUser.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
-const { logger } = require('@librechat/data-schemas');
-const { getRandomValues, hashToken } = require('~/server/utils/crypto');
+const { logger, hashToken } = require('@librechat/data-schemas');
+const { getRandomValues } = require('~/server/utils/crypto');
 const { createToken, findToken } = require('~/models');
 
 /**
