+++ packages/data-schemas/src/index.ts
@@ -1,3 +1,4 @@
+export * from './crypto';
 export { createModels } from './models';
 export { createMethods } from './methods';
 export type * from './types';
