+++ packages/data-schemas/src/schema/session.ts
@@ -1,7 +1,5 @@
 import mongoose, { Schema } from 'mongoose';
-import jwt from 'jsonwebtoken';
-import { webcrypto } from 'node:crypto';
-import { ISession, SignPayloadParams } from '~/types';
+import { ISession } from '~/types';
 
 const sessionSchema: Schema<ISession> = new Schema({
   refreshTokenHash: {
@@ -19,19 +17,5 @@ const sessionSchema: Schema<ISession> = new Schema({
     required: true,
   },
 });
-
-export async function signPayload({
-  payload,
-  secret,
-  expirationTime,
-}: SignPayloadParams): Promise<string> {
-  return jwt.sign(payload, secret!, { expiresIn: expirationTime });
-}
-
-export async function hashToken(str: string): Promise<string> {
-  const data = new TextEncoder().encode(str);
-  const hashBuffer = await webcrypto.subtle.digest('SHA-256', data);
-  return Buffer.from(hashBuffer).toString('hex');
-}
 
 export default sessionSchema;
